+++ cleo_plugins/Math/Math.cpp
@@ -5,10 +5,10 @@
 using namespace CLEO;
 using namespace plugin;
 
-class IntOperations
+class Math
 {
 public:
-    IntOperations()
+    Math()
     {
         auto cleoVer = CLEO_GetVersion();
         if (cleoVer < CLEO_VERSION)
@@ -18,11 +18,14 @@ public:
             return;
         }
 
-        //register opcodes
+        // register opcodes
         CLEO_RegisterOpcode(0x0A8E, opcode_0A8E); // x = a + b (int)
         CLEO_RegisterOpcode(0x0A8F, opcode_0A8F); // x = a - b (int)
         CLEO_RegisterOpcode(0x0A90, opcode_0A90); // x = a * b (int)
         CLEO_RegisterOpcode(0x0A91, opcode_0A91); // x = a / b (int)
+
+        CLEO_RegisterOpcode(0x0AEE, opcode_0AEE); // pow
+        CLEO_RegisterOpcode(0x0AEF, opcode_0AEF); // log
 
         CLEO_RegisterOpcode(0x0B10, Script_IntOp_AND);
         CLEO_RegisterOpcode(0x0B11, Script_IntOp_OR);
@@ -39,6 +42,12 @@ public:
         CLEO_RegisterOpcode(0x0B1C, Scr_IntOp_SHR);
         CLEO_RegisterOpcode(0x0B1D, Scr_IntOp_SHL);
         CLEO_RegisterOpcode(0x0B1E, Sign_Extend);
+
+        CLEO_RegisterOpcode(0x2700, opcode_2700); // is_bit_set
+        CLEO_RegisterOpcode(0x2701, opcode_2701); // set_bit
+        CLEO_RegisterOpcode(0x2702, opcode_2702); // clear_bit
+        CLEO_RegisterOpcode(0x2703, opcode_2703); // toggle_bit
+        CLEO_RegisterOpcode(0x2704, opcode_2704); // is_truthy
     }
 
     //0A8E=3,%3d% = %1d% + %2d% ; int
@@ -86,6 +95,30 @@ public:
         auto result = a / b;
 
         OPCODE_WRITE_PARAM_INT(result);
+        return OR_CONTINUE;
+    }
+
+    //0AEE=3,%3d% = %1d% exp %2d% // all floats
+    static OpcodeResult __stdcall opcode_0AEE(CRunningScript* thread)
+    {
+        auto base = OPCODE_READ_PARAM_FLOAT();
+        auto exponent = OPCODE_READ_PARAM_FLOAT();
+
+        auto result = (float)pow(base, exponent);
+
+        OPCODE_WRITE_PARAM_FLOAT(result);
+        return OR_CONTINUE;
+    }
+
+    //0AEF=3,%3d% = log %1d% base %2d% // all floats
+    static OpcodeResult __stdcall opcode_0AEF(CRunningScript* thread)
+    {
+        auto argument = OPCODE_READ_PARAM_FLOAT();
+        auto base = OPCODE_READ_PARAM_FLOAT();
+
+        auto exponent = log(argument) / log(base);
+
+        OPCODE_WRITE_PARAM_FLOAT(exponent);
         return OR_CONTINUE;
     }
 
@@ -310,4 +343,95 @@ public:
         
         return OR_CONTINUE;
     }
-} intOperations;
+
+    //2700=2,  is_bit_set value %1d% bit_index %2d%
+    static OpcodeResult WINAPI opcode_2700(CScriptThread* thread)
+    {
+        auto value = OPCODE_READ_PARAM_UINT();
+        auto bitIndex = OPCODE_READ_PARAM_INT();
+
+        if (bitIndex < 0 || bitIndex > 31)
+        {
+            SHOW_ERROR("Invalid '%d' bit index argument in script %s\nScript suspended.", bitIndex, ScriptInfoStr(thread).c_str());
+            return thread->Suspend();
+        }
+
+        bool result = (value >> bitIndex) & 1;
+
+        OPCODE_CONDITION_RESULT(result);
+        return OR_CONTINUE;
+    }
+
+    //2701=2,set_bit value %1d% bit_index %2d%
+    static OpcodeResult WINAPI opcode_2701(CScriptThread* thread)
+    {
+        auto value = OPCODE_READ_PARAM_OUTPUT_VAR_INT();
+        auto bitIndex = OPCODE_READ_PARAM_INT();
+
+        if (bitIndex < 0 || bitIndex > 31)
+        {
+            SHOW_ERROR("Invalid '%d' bit index argument in script %s\nScript suspended.", bitIndex, ScriptInfoStr(thread).c_str());
+            return thread->Suspend();
+        }
+
+        *value |= 1 << bitIndex;
+
+        return OR_CONTINUE;
+    }
+
+    //2702=2,clear_bit value %1d% bit_index %2d%
+    static OpcodeResult WINAPI opcode_2702(CScriptThread* thread)
+    {
+        auto value = OPCODE_READ_PARAM_OUTPUT_VAR_INT();
+        auto bitIndex = OPCODE_READ_PARAM_INT();
+
+        if (bitIndex < 0 || bitIndex > 31)
+        {
+            SHOW_ERROR("Invalid '%d' bit index argument in script %s\nScript suspended.", bitIndex, ScriptInfoStr(thread).c_str());
+            return thread->Suspend();
+        }
+
+        *value &= ~(1 << bitIndex);
+
+        return OR_CONTINUE;
+    }
+
+    //2703=3,toggle_bit value %1d% bit_index %2d% state %3d%
+    static OpcodeResult WINAPI opcode_2703(CScriptThread* thread)
+    {
+        auto value = OPCODE_READ_PARAM_OUTPUT_VAR_INT();
+        auto bitIndex = OPCODE_READ_PARAM_INT();
+        auto state = OPCODE_READ_PARAM_BOOL();
+
+        if (bitIndex < 0 || bitIndex > 31)
+        {
+            SHOW_ERROR("Invalid '%d' bit index argument in script %s\nScript suspended.", bitIndex, ScriptInfoStr(thread).c_str());
+            return thread->Suspend();
+        }
+
+        DWORD flag = 1 << bitIndex;
+        if (state)
+            *value |= flag;
+        else
+            *value &= ~flag;
+
+        return OR_CONTINUE;
+    }
+
+    //2704=1,  is_truthy value %1d%
+    static OpcodeResult WINAPI opcode_2704(CScriptThread* thread)
+    {
+        auto paramType = OPCODE_PEEK_PARAM_TYPE();
+
+        if(IsImmString(paramType) || IsVarString(paramType))
+        {
+            OPCODE_READ_PARAM_STRING_LEN(text, 1); // one character is all we need
+            OPCODE_CONDITION_RESULT(text[0] != '\0');
+            return OR_CONTINUE;
+        }
+
+        auto value = OPCODE_READ_PARAM_ANY32();
+        OPCODE_CONDITION_RESULT(value != 0);
+        return OR_CONTINUE;
+    }
+} Math;